Fixed non-explicit casts, thanks ZuluOne!
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@40 37977431-3df6-0310-b722-df95706aa16b

+++ rts/AirCAI.cpp
@@ -297,7 +297,7 @@ void CAirCAI::SlowUpdate()
 					attackPos.y=ground->GetHeight(attackPos.x,attackPos.z);
 					owner->AttackGround(attackPos,false);
 				} else {
-					int num=gs->randFloat()*eu.size();
+					int num=(int) (gs->randFloat()*eu.size());
 					orderTarget=uh->units[eu[num]];
 					owner->AttackUnit(orderTarget,false);
 					AddDeathDependence(orderTarget);
